requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
gunicorn